docs-rtd/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
accessible-pygments==0.0.5
    # via pydata-sphinx-theme
alabaster==0.7.16
    # via sphinx
babel==2.17.0
    # via
    #   pydata-sphinx-theme
    #   sphinx
beautifulsoup4==4.13.5
    # via pydata-sphinx-theme
certifi==2025.8.3
    # via requests
charset-normalizer==3.4.3
    # via requests
docutils==0.19
    # via
    #   -r requirements.in
    #   pydata-sphinx-theme
    #   sphinx
idna==3.10
    # via requests
imagesize==1.4.1
    # via sphinx
jinja2==3.1.6
    # via sphinx
markupsafe==3.0.2
    # via jinja2
packaging==25.0
    # via
    #   pydata-sphinx-theme
    #   sphinx
pydata-sphinx-theme==0.15.4
    # via sphinx-book-theme
pyenchant==3.2.2
    # via
    #   -r requirements.in
    #   sphinxcontrib-spelling
pygments==2.19.2
    # via
    #   accessible-pygments
    #   pydata-sphinx-theme
    #   sphinx
requests==2.32.5
    # via
    #   sphinx
    #   sphinxcontrib-spelling
snowballstemmer==3.0.1
    # via sphinx
soupsieve==2.8
    # via beautifulsoup4
sphinx==5.3.0
    # via
    #   -r requirements.in
    #   pydata-sphinx-theme
    #   sphinx-book-theme
    #   sphinx-copybutton
    #   sphinx-last-updated-by-git
    #   sphinxcontrib-spelling
sphinx-book-theme==1.1.3
    # via -r requirements.in
sphinx-copybutton==0.5.2
    # via -r requirements.in
sphinx-last-updated-by-git==0.3.8
    # via sphinx-sitemap
sphinx-sitemap==2.8.0
    # via -r requirements.in
sphinxcontrib-applehelp==2.0.0
    # via sphinx
sphinxcontrib-devhelp==2.0.0
    # via sphinx
sphinxcontrib-htmlhelp==2.1.0
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==2.0.0
    # via sphinx
sphinxcontrib-serializinghtml==2.0.0
    # via sphinx
sphinxcontrib-spelling==8.0.1
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   beautifulsoup4
    #   pydata-sphinx-theme
urllib3==2.5.0
    # via requests
